djangoapi/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-crispy-forms==1.9.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-nine==0.2.3
django-six==1.0.4
djangorestframework==3.11.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
httpie==2.2.0
idna==2.9
Pillow==7.1.2
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
